Final Updates 05/31/23

+++ routes/web.php
@@ -1,10 +1,11 @@
 <?php
 
-use App\Http\Controllers\ListingController;
+use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\Models\Listing;
 use Database\Factories\ListingFactory;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\ListingController;
 
 /*
 |--------------------------------------------------------------------------
@@ -20,22 +21,42 @@ use Database\Factories\ListingFactory;
 Route::get('/', [ListingController::class, 'index']);
 
 //show create form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 //Store listing data
 Route::post('/listings', [ListingController::class, 'store']);
 
 //show edit form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 //Edit to submit
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 //Delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
+
+//Manage listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 //Single listings
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
+
+//Show register/create form
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
+
+//Create New user
+Route::post('/users', [UserController::class, 'store']);
+
+//Log user out
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+//Show login form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+//Login users
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
+
+
 
 // Route::get('/hello', function () {
 //     return response('<h1>Hello world</h1>', 200)
